Implement JWT token for delete and get users

+++ src/main/java/com/cursojava/curso/controllers/UserController.java
@@ -2,10 +2,13 @@ package com.cursojava.curso.controllers;
 
 import com.cursojava.curso.dao.UserDao;
 import com.cursojava.curso.models.User;
+import com.cursojava.curso.utils.JWTUtil;
 import de.mkammerer.argon2.Argon2;
 import de.mkammerer.argon2.Argon2Factory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -13,14 +16,22 @@ public class UserController {
 
     @Autowired
     private UserDao userDao;
+    @Autowired
+    private JWTUtil jwtUtil;
 
     @RequestMapping(value = "api/users", method = RequestMethod.GET)
-    public List<User> getUsers(){
+    public List<User> getUsers(@RequestHeader(value="Authorization") String token){
+        if(!validateToken(token)){
+            return null;
+        }
         return userDao.getUsers();
     }
 
     @RequestMapping(value = "api/users/{id}", method = RequestMethod.DELETE)
-    public void deleteUser(@PathVariable Long id){
+    public void deleteUser(@PathVariable Long id, @RequestHeader(value="Authorization") String token){
+        if(!validateToken(token)){
+            return;
+        }
         userDao.delete(id);
     }
 
@@ -30,6 +41,11 @@ public class UserController {
         String hash = argon2.hash(2, 1024, 1, user.getPassword());
         user.setPassword(hash);
         userDao.create(user);
+    }
+
+    private boolean validateToken(String token){
+        String userId = jwtUtil.getKey(token);
+        return userId != null;
     }
 
 }
